requirements.txt
Flask==2.0.1
PyPDF2==1.26.0
python-docx==0.8.11
scikit-learn==0.24.2
pymysql==1.0.2
numpy==1.21.0
Werkzeug==2.0.1